requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.0
zhipuai==2.0.1 